+++ policyengine_us/variables/gov/states/nh/tax/income/exemptions/nh_old_age_exemption.py
@@ -14,16 +14,20 @@ class nh_old_age_exemption(Variable):
         p = parameters(period).gov.states.nh.tax.income.exemptions
 
         # Get the individual age.
-        age_head = tax_unit("age_head", period) 
+        age_head = tax_unit("age_head", period)
 
         # Check if the individual's eligiblity.
-        head_eligible = tax_unit.sum(age_head >= p.old_age_eligibility)
+        head_eligible = tax_unit.sum(age_head >= p.old_age_eligibility).astype(
+            int
+        )
 
         # Get the individual's spouse age.
         age_spouse = tax_unit("age_spouse", period)
 
         # Check if the individual spouse's eligiblity.
-        spouse_eligible = tax_unit.sum(age_spouse >= p.old_age_eligibility)
+        spouse_eligible = tax_unit.sum(
+            age_spouse >= p.old_age_eligibility
+        ).astype(int)
 
         # Calculate total blind exemption.
         return (head_eligible + spouse_eligible) * p.amount.old_age_addition
